Lisätty maanomistajaluettelo aloituskuulutuksen julkaisuun ja asianhallintasynkronointiin
- Lisätty MAANOMISTAJALUETTELO_ALOITUSKUULUTUS AsiakirjaTyyppi-enumiin
- Maanomistajaluettelo (Excel) generoidaan julkaisun yhteydessä
- Lisätty synkronointipyyntöön SisainenProjektiPaths:n kautta
- Poistetaan hylkäyksen yhteydessä
- Ohitetaan generointi kun uudelleenkuulutuksessa ei tiedoteta kiinteistönomistajia

AI-Assisted: Amazon Q developer, generated

+++ backend/src/handler/tila/aloitusKuulutusTilaManager.ts
@@ -23,10 +23,11 @@ import { fileService } from "../../files/fileService";
 import { parseDate } from "../../util/dateUtil";
 import { KuulutusTilaManager } from "./KuulutusTilaManager";
 import { pdfGeneratorClient } from "../../asiakirja/lambda/pdfGeneratorClient";
+import { tallennaMaanomistajaluettelo } from "../../mml/tiedotettavatExcel";
 import { IllegalAineistoStateError, IllegalArgumentError } from "hassu-common/error";
 import { projektiAdapter } from "../../projekti/adapter/projektiAdapter";
 import assert from "assert";
-import { ProjektiPaths } from "../../files/ProjektiPath";
+import { ProjektiPaths, SisainenProjektiPaths } from "../../files/ProjektiPath";
 import { ProjektiTiedostoManager } from "../../tiedostot/ProjektiTiedostoManager";
 import { requireAdmin, requireOmistaja, requirePermissionMuokkaa } from "../../user/userService";
 import { sendAloitusKuulutusApprovalMailsAndAttachments } from "../email/emailHandler";
@@ -247,6 +248,19 @@ class AloitusKuulutusTilaManager extends KuulutusTilaManager<AloitusKuulutus, Al
     aloitusKuulutusJulkaisu.muokkaaja = muokkaaja.uid;
 
     await this.generatePDFs(projekti, aloitusKuulutusJulkaisu);
+    if (
+      !aloitusKuulutusJulkaisu.uudelleenKuulutus ||
+      aloitusKuulutusJulkaisu.uudelleenKuulutus.tiedotaKiinteistonomistajia === undefined ||
+      aloitusKuulutusJulkaisu.uudelleenKuulutus.tiedotaKiinteistonomistajia
+    ) {
+      aloitusKuulutusJulkaisu.maanomistajaluettelo = await tallennaMaanomistajaluettelo(
+        projekti,
+        new SisainenProjektiPaths(projekti.oid).aloituskuulutus(aloitusKuulutusJulkaisu),
+        this.vaihe,
+        aloitusKuulutusJulkaisu.kuulutusPaiva,
+        aloitusKuulutusJulkaisu.id
+      );
+    }
     await projektiDatabase.aloitusKuulutusJulkaisut.insert(projekti.oid, aloitusKuulutusJulkaisu);
     await approvalEmailSender.sendEmails(projekti, tilasiirtymaTyyppi);
   }
@@ -261,6 +275,13 @@ class AloitusKuulutusTilaManager extends KuulutusTilaManager<AloitusKuulutus, Al
     aloitusKuulutus.palautusSyy = syy;
     if (julkaisuWaitingForApproval.aloituskuulutusPDFt) {
       await this.deletePDFs(projekti.oid, julkaisuWaitingForApproval.aloituskuulutusPDFt);
+    }
+    if (julkaisuWaitingForApproval.maanomistajaluettelo) {
+      await fileService.deleteYllapitoFileFromProjekti({
+        oid: projekti.oid,
+        filePathInProjekti: julkaisuWaitingForApproval.maanomistajaluettelo,
+        reason: "Aloituskuulutus rejected",
+      });
     }
     await projektiDatabase.saveProjekti({ oid: projekti.oid, versio: projekti.versio, aloitusKuulutus });
     await projektiDatabase.aloitusKuulutusJulkaisut.delete(projekti, julkaisuWaitingForApproval.id);
@@ -321,11 +342,13 @@ class AloitusKuulutusTilaManager extends KuulutusTilaManager<AloitusKuulutus, Al
           filePathInProjekti: pdfs.aloituskuulutusIlmoitusPDFPath,
           reason: "Aloituskuulutus rejected",
         });
-        await fileService.deleteYllapitoFileFromProjekti({
-          oid,
-          filePathInProjekti: pdfs.aloituskuulutusIlmoitusKiinteistonOmistajallePDFPath,
-          reason: "Aloituskuulutus rejected",
-        });
+        if (pdfs.aloituskuulutusIlmoitusKiinteistonOmistajallePDFPath) {
+          await fileService.deleteYllapitoFileFromProjekti({
+            oid,
+            filePathInProjekti: pdfs.aloituskuulutusIlmoitusKiinteistonOmistajallePDFPath,
+            reason: "Aloituskuulutus rejected",
+          });
+        }
       }
     }
   }
